Bitmap.cpp, Data.cpp: * changed data size member to s64, added cast when used with stb_image
Texture:
* disabled glContext usage for now to check compile
* disabled GLEE usage
* removed obsolete casts
* removed obsolete bitmapComponent conversion
* replaced GLDEBUG_THROW with GLASSERT
* removed old namespaces

lgl:
* added return value that should never be needed because compiler complained and didn't recognize abort() call in ASSERT macro
* added missing lost namespace in header
* fixed ASSERT macro usage and added empty ASSERT definition for disabled GL debugging

prefix-shared.h:
* cleaned up
* added type includes
* added lgl include

types.h:
* added 64 bit ints

xcode:
* added new prefix headers for debug and release build

+++ lost/Bitmap.cpp
@@ -193,7 +193,7 @@ void Bitmap::init(const DataPtr& inData)
   destroy();
 //  DOUT("init image from memory: " << inFile->location);
   int bytesPerPixel, w, h;
-  data = stbi_load_from_memory(reinterpret_cast<const stbi_uc*>(inData->bytes.get()), inData->size, &w, &h, &bytesPerPixel, 0);
+  data = stbi_load_from_memory(reinterpret_cast<const stbi_uc*>(inData->bytes.get()), (s32)inData->size, &w, &h, &bytesPerPixel, 0);
   width = w;
   height = h;
   ASSERT(data, "couldn't init image from memory: "<<std::string(stbi_failure_reason()));
